Use Git commit dates for "Last modified" dates of issues
The makefile target "dates" will create a file containing the git commit
timestamp of each xml/issue*.xml file. That file gets read and used to
set the issue::mod_time data member to the XML file's commit date
instead of its filesystem::last_write_time().

This means the issues list shows meaningful and reproducable "Last
modified" dates, instead of the time when the file happened to be
touched (which is usually when the git branch was checked out, not when
the issue was edited).