Sort section-data output properly
The current code only sorts on the section numbers, so the tags within a
section are output in an unspecified order. This means regenerating the
section-data file can produce spurious diffs like:

     6.1 [basic.pre]
     6.2 [basic.def]
-    6.3 [term.odr.use]
     6.3 [basic.def.odr]
+    6.3 [term.odr.use]
     6.4 [basic.scope]
         6.4.1 [basic.scope.scope]
         6.4.2 [basic.scope.pdecl]

Use both parts of the pair when sorting, so that the output is
predictable and doesn't have these pointless changes.